Add scripts to aid benchmarking
Use `run_test.sh` to define which tests to run and in which order and
store the execution results, including the results of `time` command.

`total_seq_ms.sh` processes a logfile from an execution and aggregates
all the duration messages to estimate the runtime of a purely
sequential implementation.