Use []DataObject in SplitInBatches
`DataObject` includes an `ExternalID` field that is carried through
the classification, but access to this field was not available if
`SplitInBatches` was used.

Expose the `DataObject` struct directly in the signature of
`SplitInBatches` so users have access to that field. They will need to
serialize their input into an array of `DataObjects` instead of an
array of `string`s.

Signed-off-by: Miguel Bernabeu <miguel.bernabeu@devex.com>