Allow client configuration of underlying settings
This allows to set a custom HTTP clientor transport layer, as needed,
and also to direct your requests towards a different server than the
official API (to provide a custom proxy, mocking, or any reason).

To keep the ease of use, introduce `NewDefaultClient` with the same
interface and default values to be a drop-in replacement of the
previous `NewClient` method.

Signed-off-by: Miguel Bernabeu <miguel.bernabeu@devex.com>